move all the banner code
