+++ desktop/banner_ctrl.js
@@ -16,11 +16,6 @@ import CampaignDays, { startOfDay, endOfDay } from '../shared/campaign_days';
 import InterruptibleTimeout from '../shared/interruptible_timeout';
 import Translations from '../shared/messages/de';
 
-const Handlebars = require( 'handlebars/runtime' );
-Handlebars.registerHelper( 'capitalizeFirstLetter', function ( message ) {
-	return message.charAt( 0 ).toUpperCase() + message.slice( 1 );
-} );
-
 const DevGlobalBannerSettings = require( '../shared/global_banner_settings' );
 const GlobalBannerSettings = window.GlobalBannerSettings || DevGlobalBannerSettings;
 const BannerFunctions = require( '../shared/banner_functions' )( GlobalBannerSettings, Translations );
@@ -49,32 +44,28 @@ const $bannerContainer = $( '#WMDE-Banner-Container' );
 const CampaignName = $bannerContainer.data( 'campaign-tracking' );
 const BannerName = $bannerContainer.data( 'tracking' );
 const sizeIssueIndicator = new SizeIssueIndicator( sizeIssueThreshold );
+const ProgressBar = require( '../shared/progress_bar/progress_bar' );
+const progressBar = new ProgressBar( GlobalBannerSettings, campaignProjection, {
+	textRight: '',
+	textInnerRight: 'Geschafft: <span class="js-donation_value">0,0</span> Mio. Euro'
+} );
 const bannerDisplayTimeout = new InterruptibleTimeout();
 
 $bannerContainer.html( bannerTemplate( {
 	numberOfDonors: donorFormatter( campaignProjection.getProjectedNumberOfDonors() ),
 	CampaignName: CampaignName,
 	BannerName: BannerName,
+	progressBar: progressBar.render(),
 	numberOfNewMembersLastYear: '64.334', // @todo Update with real number, or add to projections,
 	becomeMemberLink: 'https://spenden.wikimedia.de/apply-for-membership?skin=10h16&type=sustaining&' +
 		'piwik_campaign=' + CampaignName + '&piwik_kwd=' + BannerName
 } ) );
 
-// BEGIN form init code
-
 const trackingEvents = new TrackingEvents( trackingBaseUrl, BannerName, $( '.click-tracking__pixel' ) );
 
-BannerFunctions.initializeBannerEvents();
-
-// END form init code
-
-$( '.show-beef', $bannerContainer ).click( function () {
-	$( '.banner', $bannerContainer ).toggleClass( 'enabled' );
-	addSpaceInstantly();
-} );
-
-$( '.close__beef', $bannerContainer ).click( function () {
-	$( '.banner', $bannerContainer ).removeClass( 'enabled' );
+$( '.expand_button' ).on( 'click', function ( e ) {
+	e.preventDefault();
+	$( 'div#WMDE_Banner .banner' ).toggleClass( 'expanded' );
 	addSpaceInstantly();
 } );
 
@@ -110,13 +101,25 @@ function displayBanner() {
 	bannerElement.animate( { top: 0 }, 1000 );
 	setTimeout( function () { progressBar.animate(); }, 1000 );
 
+	var $numberOfDonors = $( '.infobox .mobile-only .number_of_donors' );
+	$( { Counter: 0 } ).animate(
+		{ Counter: campaignProjection.getProjectedNumberOfDonors() },
+		{
+			duration: 3000,
+			easing: 'swing',
+			step: function () {
+				$numberOfDonors.text( donorFormatter( Math.ceil( this.Counter ) ) );
+			}
+		}
+	);
+
 	$( window ).resize( function () {
 		addSpaceInstantly();
 	} );
 }
 
 // track banner expansion and closing
-trackingEvents.trackClickEvent( $( 'thankyou-message__call-to-action' ), 'banner-expanded', 1 );
+trackingEvents.trackClickEvent( $( '.expand_button' ), 'banner-expanded', 1 );
 trackingEvents.trackClickEvent( $( '#WMDE_Banner .close__link' ), 'banner-closed', bannerCloseTrackRatio );
 
 // BEGIN Banner close functions
@@ -135,8 +138,6 @@ $( '#ca-ve-edit, .mw-editsection-visualeditor' ).click( function () {
 	$( '#WMDE_Banner' ).hide();
 	removeBannerSpace();
 } );
-
-// END Banner close functions
 
 // Display banner on load
 $( function () {
