+++ desktop/banner_s1.js
@@ -1,9 +1,6 @@
-require( './css/styles.pcss' );
+require( './css/styles_s1.pcss' );
 require( './css/icons.css' );
 require( './css/wlightbox.css' );
-
-// For A/B testing different styles, load
-// require( './css/styles_var.pcss' );
 
 // BEGIN Banner-Specific configuration
 const bannerCloseTrackRatio = 0.01;
@@ -19,6 +16,11 @@ import CampaignDays, { startOfDay, endOfDay } from '../shared/campaign_days';
 import InterruptibleTimeout from '../shared/interruptible_timeout';
 import Translations from '../shared/messages/de';
 
+const Handlebars = require( 'handlebars/runtime' );
+Handlebars.registerHelper( 'capitalizeFirstLetter', function ( message ) {
+	return message.charAt( 0 ).toUpperCase() + message.slice( 1 );
+} );
+
 const DevGlobalBannerSettings = require( '../shared/global_banner_settings' );
 const GlobalBannerSettings = window.GlobalBannerSettings || DevGlobalBannerSettings;
 const BannerFunctions = require( '../shared/banner_functions' )( GlobalBannerSettings, Translations );
@@ -32,13 +34,14 @@ const campaignProjection = new CampaignProjection(
 		baseDonationSum: GlobalBannerSettings[ 'donations-collected-base' ],
 		donationAmountPerMinute: GlobalBannerSettings[ 'appr-donations-per-minute' ],
 		donorsBase: GlobalBannerSettings[ 'donators-base' ],
-		donorsPerMinute: GlobalBannerSettings[ 'appr-donators-per-minute' ]
+		donorsPerMinute: GlobalBannerSettings[ 'appr-donators-per-minute' ],
+		goalDonationSum: GlobalBannerSettings.goalSum
 	}
 );
 const formatNumber = require( 'format-number' );
 const donorFormatter = formatNumber( { round: 0, integerSeparator: '.' } );
 
-const bannerTemplate = require( './templates/banner_html.hbs' );
+const bannerTemplate = require( './templates/banner_html_s1.hbs' );
 
 const $ = require( 'jquery' );
 
@@ -46,28 +49,32 @@ const $bannerContainer = $( '#WMDE-Banner-Container' );
 const CampaignName = $bannerContainer.data( 'campaign-tracking' );
 const BannerName = $bannerContainer.data( 'tracking' );
 const sizeIssueIndicator = new SizeIssueIndicator( sizeIssueThreshold );
-const ProgressBar = require( '../shared/progress_bar/progress_bar' );
-const progressBar = new ProgressBar( GlobalBannerSettings, campaignProjection, {
-	textRight: '',
-	textInnerRight: 'Geschafft: <span class="js-donation_value">0,0</span> Mio. Euro'
-} );
 const bannerDisplayTimeout = new InterruptibleTimeout();
 
 $bannerContainer.html( bannerTemplate( {
 	numberOfDonors: donorFormatter( campaignProjection.getProjectedNumberOfDonors() ),
 	CampaignName: CampaignName,
 	BannerName: BannerName,
-	progressBar: progressBar.render(),
 	numberOfNewMembersLastYear: '64.334', // @todo Update with real number, or add to projections,
-	becomeMemberLink: 'https://spenden.wikimedia.de/apply-for-membership?skin=cat17&type=sustaining&' +
-	'piwik_campaign=' + CampaignName + '&piwik_kwd=' + BannerName
+	becomeMemberLink: 'https://spenden.wikimedia.de/apply-for-membership?skin=10h16&type=sustaining&' +
+		'piwik_campaign=' + CampaignName + '&piwik_kwd=' + BannerName
 } ) );
+
+// BEGIN form init code
 
 const trackingEvents = new TrackingEvents( trackingBaseUrl, BannerName, $( '.click-tracking__pixel' ) );
 
-$( '.expand_button' ).on( 'click', function ( e ) {
-	e.preventDefault();
-	$( 'div#WMDE_Banner .banner' ).toggleClass( 'expanded' );
+BannerFunctions.initializeBannerEvents();
+
+// END form init code
+
+$( '.show-beef', $bannerContainer ).click( function () {
+	$( '.banner', $bannerContainer ).toggleClass( 'enabled' );
+	addSpaceInstantly();
+} );
+
+$( '.close__beef', $bannerContainer ).click( function () {
+	$( '.banner', $bannerContainer ).removeClass( 'enabled' );
 	addSpaceInstantly();
 } );
 
@@ -103,25 +110,13 @@ function displayBanner() {
 	bannerElement.animate( { top: 0 }, 1000 );
 	setTimeout( function () { progressBar.animate(); }, 1000 );
 
-	var $numberOfDonors = $( '.infobox .mobile-only .number_of_donors' );
-	$( { Counter: 0 } ).animate(
-		{ Counter: campaignProjection.getProjectedNumberOfDonors() },
-		{
-			duration: 3000,
-			easing: 'swing',
-			step: function () {
-				$numberOfDonors.text( donorFormatter( Math.ceil( this.Counter ) ) );
-			}
-		}
-	);
-
 	$( window ).resize( function () {
 		addSpaceInstantly();
 	} );
 }
 
 // track banner expansion and closing
-trackingEvents.trackClickEvent( $( '.expand_button' ), 'banner-expanded', 1 );
+trackingEvents.trackClickEvent( $( 'thankyou-message__call-to-action' ), 'banner-expanded', 1 );
 trackingEvents.trackClickEvent( $( '#WMDE_Banner .close__link' ), 'banner-closed', bannerCloseTrackRatio );
 
 // BEGIN Banner close functions
@@ -140,6 +135,8 @@ $( '#ca-ve-edit, .mw-editsection-visualeditor' ).click( function () {
 	$( '#WMDE_Banner' ).hide();
 	removeBannerSpace();
 } );
+
+// END Banner close functions
 
 // Display banner on load
 $( function () {
