--- desktop/css/styles_var.pcss
@@ -1,302 +1,0 @@
-#WMDE_Banner {
-
-	/**
-	 * GLOBAL RESET
-	 */
-	img {
-		border: 0 none;
-	}
-
-	/**
-	 * CUSTOM
-	 */
-
-	display: none;
-	width: 100%;
-	position: absolute;
-	z-index: 10000;
-	font-size: 16px;
-
-	.progress_bar__donation_fill {
-		height: 40px;
-		background-color: #009ee0;
-		border-radius: 8px;
-	}
-
-	.progress_bar__donation_text {
-		line-height: 40px;
-		font-family: 'Helvetica', 'Verdana', sans-serif;
-		font-size: 16px;
-	}
-
-	.progress_bar__wrapper {
-		border: 3px solid #ffffff;
-		margin: 0;
-	}
-
-	.progress_bar__donation_remaining--inner {
-		padding-right: 0;
-	}
-
-	.text__paragraph {
-		margin-bottom: 20px;
-		line-height: 140%;
-	}
-
-	.text__paragraph--listintro,
-	.text__paragraph--closing {
-		margin-bottom: 0;
-	}
-
-	.text__headline--bold,
-	.text__paragraph--bold {
-		font-weight: bold;
-	}
-
-	.text__headline--italic,
-	.text__paragraph--italic {
-		font-style: italic;
-	}
-
-	.teaser {
-		background-color: #dce4e8;
-		margin: 10px 10px 20px 10px;
-		padding: 20px 20px 0 20px;
-		height: 180px;
-		position: relative;
-
-		.infobox__text {
-			color: #000000;
-			font-size: 16px;
-			line-height: 1.2em;
-			text-align: left;
-			padding-top: 10px;
-
-			display: flex;
-			flex-direction: row;
-			align-items: flex-start;
-
-			.text__paragraph {
-				flex: 1 1 auto;
-				width: 50%;
-				padding: 0 20px;
-			}
-		}
-
-		.expand_button {
-			display: block;
-			position: absolute;
-			bottom: 0;
-			right: 1.5em;
-
-			color: #ffffff;
-			font-family: 'Helvetica', 'Verdana', sans-serif;
-			font-size: 16px;
-			background-color: #0065a4;
-			line-height: 60px;
-			height: 60px;
-			width: 215px;
-			text-align: center;
-		}
-
-		.expand_button__text--collapsed,
-		.expand_button__text--expanded {
-			padding-right: 36px;
-			background-size: 36px 36px;
-			background-repeat: no-repeat;
-			background-position: right center;
-		}
-
-		.expand_button__text--collapsed {
-			display: inline;
-			background-image: svg-load( '../assets/Ic_expand_more_36px.svg', fill=#ffffff );
-		}
-
-		.expand_button__text--expanded {
-			display: none;
-			background-image: svg-load( '../assets/Ic_expand_less_36px.svg', fill=#ffffff );
-		}
-	}
-
-	.beef {
-		display: none;
-
-		margin: 15px 30px 20px 30px;
-		color: #ffffff;
-
-		.beef__column {
-			flex: 1 1 auto;
-			width: 50%;
-			margin-bottom: 20px;
-			padding: 0 20px;
-		}
-
-		.feature_list {
-			margin-bottom: 20px;
-			list-style: disc none;
-		}
-
-		.become_member__button {
-			display: inline-block;
-			width: 280px;
-			height: 40px;
-			line-height: 40px;
-			background-color: #009ee0;
-			color: #ffffff;
-			font-family: 'Helvetica', 'Verdana', sans-serif;
-			font-size: 16px;
-			font-weight: bold;
-			text-align: center;
-			margin: 20px 0;
-		}
-	}
-
-	.banner {
-		font-family: 'Linux Libertine', 'Linux Libertine O', 'Georgia', 'Bitstream Charter', serif;
-		box-shadow: 0 3px 5px rgba( 0, 0, 0, 0.5 );
-		background-color: #0065a4;
-
-		display: flex;
-		flex-direction: column;
-	}
-
-	.banner.expanded {
-		.teaser {
-			.expand_button__text--collapsed {
-				display: none;
-			}
-
-			.expand_button__text--expanded {
-				display: inline;
-			}
-		}
-
-		.beef {
-			display: flex;
-		}
-	}
-
-	.close {
-		position: absolute;
-		width: 18px;
-		margin: 0.25em 0.25em 0.25em 0;
-		top: 0;
-		right: 0;
-		flex: auto;
-	}
-
-	.close img {
-		width: 14px;
-		height: 14px;
-	}
-
-	.click-tracking {
-		display: none;
-	}
-
-	.click-tracking__pixel {
-		height: 0;
-		width: 0;
-		border: 0 none;
-	}
-
-	@media screen and ( max-width: 960px ) {
-		.text__paragraph {
-			margin-bottom: 10px;
-			font-size: 12px;
-		}
-
-		.progress_bar__donation_text {
-			font-size: 12px;
-		}
-
-		.teaser {
-			height: 160px;
-			margin: 5px 5px 10px 5px;
-
-			.expand_button {
-				font-size: 12px;
-				height: 50px;
-				width: 160px;
-			}
-		}
-
-		.beef {
-			.feature_list {
-				font-size: 12px;
-			}
-
-			.become_member__button {
-				font-size: 12px;
-				width: 230px;
-				height: 30px;
-				line-height: 30px;
-			}
-		}
-	}
-
-	@media screen and ( max-width: 660px ) {
-		.desktop-only {
-			display: none !important;
-		}
-
-		.text__paragraph {
-			font-size: 18px;
-		}
-
-		.teaser {
-			height: 190px;
-			padding-top: 10px;
-			margin: 6px 6px 12px 6px;
-
-			.expand_button {
-				margin: auto;
-				left: 0;
-				right: 0;
-			}
-
-			.infobox__text {
-				flex-direction: column;
-				padding-top: 0;
-
-				.text__paragraph {
-					width: 100%;
-					padding: 0;
-					margin: 0;
-					text-align: center;
-					font-size: 18px;
-				}
-
-				.number_of_donors {
-					font-size: 30px;
-				}
-			}
-		}
-
-		.beef {
-			flex-direction: column;
-
-			.beef__column {
-				width: 100%;
-				padding: 0;
-			}
-
-			.feature_list {
-				font-size: 18px;
-			}
-
-			.become_member__button {
-				font-size: 18px;
-				width: 300px;
-				height: 40px;
-				line-height: 40px;
-			}
-		}
-	}
-
-	@media screen and ( min-width: 661px ) {
-		.mobile-only {
-			display: none !important;
-		}
-	}
-}
