+++ desktop/templates/banner_html_s1.hbs
@@ -1,0 +1,87 @@
+<div id="WMDE_Banner">
+	<div class="banner">
+		<div class="teaser">
+			<div class="thankyou-message">
+				<div class="thankyou-message__image">
+					<img src="https://upload.wikimedia.org/wikipedia/commons/a/ac/Thank-you-2017-text-thanks.svg">
+				</div>
+
+				<div class="thankyou-message__text">für <span class="thankyou-message__text--num-donors">{{ numberOfDonors }}</span> Spenden in sieben Wochen!</div>
+				<div class="thankyou-message__call-to-action">
+					<a href="#" class="show-beef">
+						<span class="thankyou-message__show-text">Dankesbotschaft anzeigen</span>
+						<span class="thankyou-message__hide-text">Dankesbotschaft verbergen</span>
+					</a>
+				</div>
+			</div>
+
+			<div class="author-message">
+				<div class="author-message__text">
+					Spenden sind eine Möglichkeit, sich für Freies Wissen zu engagieren &ndash; wussten Sie,
+					dass Sie bei Wikipedia auch mitmachen können?
+				</div>
+				<div class="author-message__call-to-action desktop-only">
+					<a href="https://wikipedia.de/lernewikipedia?campaign=wmde_etc2017_bt1">Lernen, bei Wikipedia mitzuschreiben</a>
+				</div>
+				<div class="author-message__call-to-action mobile-only">
+					<a href="https://wikipedia.de/lernewikipedia?campaign=wmde_etc2017_bt1">Jetzt Bearbeiten lernen</a>
+				</div>
+			</div>
+		</div>
+		<div class="beef">
+			<p class="text__paragraph">Liebe Leserinnen und Leser,</p>
+
+			<p class="text__paragraph">Wikipedia ist werbefrei. Aber einmal im Jahr rufen wir ein paar Wochen lang
+			zur Unterstützung von Wikipedia auf. Die Resonanz der hunderttausenden
+			Menschen ist überwältigend und zeigt, wie wichtig Wikipedia für so viele
+			ist.</p>
+
+			<p class="text__paragraph">Der Dank gebührt aber in erster Linie denjenigen, die das Rückgrat von
+			Wikipedia bilden: den freiwilligen Autorinnen und Autoren, die Wikipedia
+			schreiben. Das tun sie ehrenamtlich in ihrer Freizeit. Manche sogar mehrere
+			Stunden am Tag. Diese unermüdliche Arbeit an Wikipedia ist es, die
+			essentiell für die Aktualität und damit auch für die Beliebtheit bei
+			Lesenden ist.</p>
+
+			<p class="text__paragraph">Auch wenn viele den Eindruck haben, dass ja schon alles in Wikipedia steht,
+			so verändert sich die Welt und damit auch die Informationen. Deswegen
+			möchte ich Sie ermutigen, vielleicht selbst einmal etwas in einem
+			Wikipedia-Artikel zu ergänzen oder zu aktualisieren. Das ist gar nicht so
+			schwer, wie Sie vielleicht denken. Ich selbst habe vor über 10 Jahren meine
+			ersten Schritte gemacht und finde es wunderbar, selbst etwas beizutragen.
+			Wenn Sie sich das anschauen möchten, klicken sie doch auf die Seite
+			<a href="https://wikipedia.de/lernewikipedia?campaign=wmde_etc2017_bt1">wikipedia.de/lernewikipedia</a>,
+				auf der Sie selbst lernen können, Wikipedia zu
+			bearbeiten.</p>
+
+			<p class="text__paragraph">Sie sehen: Es gibt viele Möglichkeiten, sich zu beteiligen. Egal, ob mit
+			einer Spende oder mit Zeit und Wissen – es ist eine großartige Gemeinschaft
+			von Menschen, die gemeinsam der Vision des freien Zugangs zu Wissen für
+			alle Menschen auf dieser Welt jeden Tag ein Stück näher kommen.</p>
+
+			<p class="text__paragraph">Noch einmal meinen herzlichen Dank an alle, die uns unterstützen.</p>
+
+			<p class="text__paragraph">Ich wünsche Ihnen ein gesundes und erfolgreiches Jahr 2018.</p>
+
+			<p class="text__paragraph--signature"><img src="https://upload.wikimedia.org/wikipedia/commons/2/25/Signature-tim-m-hector-white.png"></p>
+			<p class="text__paragraph">Tim Moritz Hector</p>
+
+			<p class="text__paragraph">Vorsitzender des Präsidiums von Wikimedia Deutschland e. V.</p>
+			<div class="close">
+				<a href="#" class="close__beef">
+					<img src="https://upload.wikimedia.org/wikipedia/commons/4/4c/Grey_close_x.svg" alt="schließen" title="schließen">
+				</a>
+			</div>
+		</div>
+
+		<div class="close">
+			<a href="#" class="close__link">
+				<img src="https://upload.wikimedia.org/wikipedia/commons/c/ce/Black_close_x.svg" alt="schließen" title="schließen">
+			</a>
+		</div>
+	</div>
+
+	<div class="click-tracking">
+		<img class="click-tracking__pixel" src="" />
+	</div>
+</div>
