--- desktop/templates/banner_html_var.hbs
@@ -1,119 +1,0 @@
-<div id="WMDE_Banner">
-	<div class="banner">
-		<div class="teaser">
-			<div class="desktop-only">
-				{{{ progressBar }}}
-			</div>
-
-			<div class="infobox">
-				<div class="infobox__text desktop-only">
-					<p class="text__paragraph">
-						{{ numberOfDonors }} Menschen haben in den letzten Wochen gezeigt, dass Wikipedia einzigartig ist.
-						Wir bedanken uns ganz herzlich bei allen Spenderinnen und Spendern für die Unterstützung im letzten Jahr.
-					</p>
-
-					<p class="text__paragraph">
-						Doch Wikipedia ist nicht nur deshalb etwas Besonderes. Lesen Sie, warum Wikipedia mehr ist als eine Website.
-					</p>
-				</div>
-
-				<div class="infobox__text mobile-only">
-					<p class="text__paragraph number_of_donors">{{ numberOfDonors }}</p>
-					<p class="text__paragraph">
-						Menschen haben in den letzten Wochen gezeigt, dass Wikipedia einzigartig ist. Vielen Dank!
-					</p>
-				</div>
-			</div>
-
-			<a href="#" class="expand_button">
-				<span class="expand_button__text--collapsed">Mehr lesen</span>
-				<span class="expand_button__text--expanded">Weniger lesen</span>
-			</a>
-
-			<div class="close">
-				<a href="#" class="close__link">
-					<img src="//upload.wikimedia.org/wikipedia/commons/c/cb/Close_x_-_black.png" alt="schließen" title="schließen">
-				</a>
-			</div>
-		</div>
-
-		<div class="beef">
-			<div class="beef__column">
-				<p class="text__paragraph">Wikipedia ist aus verschiedenen Gründen einzigartig.</p>
-
-				<p class="text__paragraph text__paragraph--listintro">Sie ist einzigartig, weil</p>
-
-				<ul class="feature_list">
-					<li>sie als einzige Top-10-Website werbefrei und gemeinnützig betrieben wird.</li>
-					<li>Menschen in Deutschland und weltweit jeden Tag aus freien Stücken Wissen teilen und Wikipedia schreiben.</li>
-					<li>Menschen freiwillig Geld für dieses Projekt spenden und somit die laufenden Kosten und die Projekte für mehr Freies Wissen finanzieren.</li>
-					<li>dieses Freiwilligenprojekt ein nobles Ziel hat: Allen Menschen freien Zugang zu Wissen zur freien Verfügung zu stellen.</li>
-				</ul>
-
-				<p class="text__paragraph text__paragraph--bold">Kurzum: Wikipedia ist einzigartig, weil Menschen dieses Projekt ehrenamtlich erschaffen und mit Spenden möglich machen.</p>
-
-				<p class="text__paragraph desktop-only">
-					Wenngleich Wikipedia frei nutzbar ist, so entstehen doch Kosten. Kosten für die technische Infrastruktur,
-					die Weiterentwicklung der Software, die Unterstützung der Freiwilligen und die weitergehenden Projekte zur
-					Förderung Freien Wissens. Die Kosten entstehen dabei nicht nur in Deutschland, sondern auch international,
-					da Wikipedia ein internationales Projekt ist; mit über 280 unterschiedlichen Sprachversionen,
-					die jeden Monat 15 Milliarden Mal aufgerufen werden.
-				</p>
-
-				<p class="text__paragraph desktop-only">
-					Unsere jährliche Spendenkampagne finanziert diese Kosten. Wir sind sehr dankbar für das Engagement der
-					{{ numberOfDonors }} Menschen in Deutschland, die sich in den letzten Wochen für Freies Wissen eingesetzt haben.
-				</p>
-
-				<p class="text__paragraph mobile-only">
-					Wenngleich Wikipedia frei nutzbar ist, so entstehen Kosten für eine Website, die 280 unterschiedliche
-					Sprachversionen hat und jeden Monat 15 Milliarden Mal aufgerufen wird. Unsere jährliche Spendenkampagne
-					finanziert diese Kosten. Ich bin sehr dankbar für das Engagement der {{ numberOfDonors }} Menschen
-					in Deutschland, die sich in den letzten Wochen für Freies Wissen eingesetzt haben.
-				</p>
-
-				<p class="text__paragraph text__paragraph--bold">Bei Ihnen bedanke ich mich ganz herzlich.</p>
-			</div>
-
-			<div class="beef__column">
-				<p class="text__paragraph">
-					Wikipedia wird auch in Zukunft auf ihre Leserinnen und Leser angewiesen sein.
-					Daher fragen wir jedes Jahr aufs neue Menschen wie Sie nach Ihrer Unterstützung.
-					Sie können uns helfen, diese Abhängigkeit von der jährlichen Spendenkampagne zu reduzieren: Durch eine Fördermitgliedschaft.
-				</p>
-
-				<p class="text__paragraph">
-					Als Fördermitglied unterstützen Sie Freies Wissen mit einem regelmäßigen Beitrag.
-					Der Mindestbetrag für Privatpersonen beträgt jährlich 24 Euro – das sind nur 2 Euro pro Monat.
-					Der Beitrag ist dabei steuerlich absetzbar. Eine jährliche Zuwendungsbestätigung wird Ihnen automatisch zugesendet.
-					Die Fördermitgliedschaft ist jederzeit formlos per E-Mail kündbar.
-				</p>
-
-				<p class="text__paragraph">
-					{{ numberOfNewMembersLastYear }} Menschen in Deutschland haben sich in den letzten Jahren für eine Mitgliedschaft entschlossen.
-					Seien auch Sie dabei, und werden Sie Teil dieser einzigartigen Gemeinschaft, die sich regelmäßig und nachhaltig
-					für Freies Wissen in Deutschland einsetzt.
-				</p>
-
-				<a href="{{ becomeMemberLink }}" class="become_member__button">Jetzt Fördermitglied werden</a>
-
-				<p class="text__paragraph desktop-only">
-					Unser Ziel ist eine Welt, in der alle Menschen gleichermaßen ungehinderten Zugang zu Freiem Wissen haben.
-					Das schaffen wir nur gemeinsam mit Ihnen. Ich bin sehr dankbar, dass so viele Menschen ‒
-					sei es durch die aktive Mitarbeit in der Wikipedia, durch eine Spende oder durch eine Mitgliedschaft
-					bei Wikimedia Deutschland ‒ sich an der Erreichung dieses Ziels beteiligen.
-				</p>
-
-				<p class="text__paragraph text__paragraph--closing">Ich wünsche Ihnen ein gesundes und erfolgreiches Jahr 2018.</p>
-
-				<p class="text__paragraph text__paragraph--closing">Abraham Taherivand</p>
-
-				<p class="text__paragraph text__paragraph--closing">Geschäftsführender Vorstand Wikimedia Deutschland e. V.</p>
-			</div>
-		</div>
-	</div>
-
-	<div class="click-tracking">
-		<img class="click-tracking__pixel" src="" />
-	</div>
-</div>
